Add npm scripts that make it easier to generate test data
The intention is to make it easier for devs on the Eolas/Synapse project to reset the database to a known state. This is (hopefully) a step towards better, more automated integration testing.

I updated the README with instructions, so take a look there, but here's how it works:

Running `npm run generate-test-data-local` simply runs all scripts in the "test_data" folder.

Running `npm run generate-test-data-staging` does the following:
1) Reads the development.json config.
2) Modify the config to point to the Staging database.
3) Run all the scripts in "test_data".
4) Reset development.json to it's original state.